Allowance to selected customers
You are an employee of ‘Rox Travel’ channel.The channel has decided to give allowances to some customer who satisfy these conditions. The conditions are:

The customer should be born on or before july 22 1987

The month of D.O.B month should be of 31 days.

You are given with the D.O.B of all the employees.Your task is to print the employee index who are having chance to avail special offer.

Input Description:
First line contains the number of employee.Next line contains an array of D.O.B of employees

Output Description:
Print the employee index (index at 1). Print-1 if there are no such employee

Sample Input :
Input
4
23 MARCH 1996 23 MARCH 1986 22 JULY 1987 23 APRIL 1987
Sample Output :
2 3