Check if A[i] occurs A[i+1] times in the array
J.S.Bhayana gave an integer array A of size N. If A[i] occurs A[i + 1] times in the array, i is called a beautiful index. Find the number of beautiful indices in the array. 
(1 <= i < N)
(1 based indexing)

Input Description:
The first line contains an integer N, denoting the size of the array. The next line contains N space separated integers.

Output Description:
Print a line containing the number of beautiful indices in the array.

Sample Input :
4
1 2 1 4 1 1
Sample Output :
3