no of notes required to make the amount up in most optimal way
Given an amount A(integer) find the number of currency notes required to make the amount up in the most optimal way. The available denominations are 1000,500,100,50,10,1. The question is set in PRE-DEMONITISATION era so 1000 rupee notes are available :)
eg: rs 689 (* currency notes are 500,100,50,10,1 )
500+100+50+(3*10)+(9*1)
count=15
Input Size : N <= 100000000
Sample Testcase :
INPUT
689
OUTPUT
15